+++ CHANGELOG.md
@@ -3,6 +3,22 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v5.8.1
+------------------------------
+*December 21, 2021*
+
+### Updated
+- Circle CI cached components to include new f-filter-pill component
+
+
+v5.8.0
+------------------------------
+*December 20, 2021*
+
+### Updated
+- fozzie package to the latest (v6.2.0).
+
+
 v5.7.0
 ------------------------------
 *December 16, 2021*